WIP: fix transitions, messages, hud, defeat state, etc.
