+++ app/render/renderMessage.ts
@@ -1,7 +1,8 @@
-import { CANVAS_WIDTH, CANVAS_HEIGHT, GAME_KEY } from 'app/gameConstants';
-import { drawFrame } from 'app/utils/animations';
-import { fillRect, pad } from 'app/utils/index';
-import { simpleWhiteFont, keyboardMap, xboxMap } from 'app/utils/simpleWhiteFont';
+import {CANVAS_WIDTH, CANVAS_HEIGHT, GAME_KEY} from 'app/gameConstants';
+import type {MessageScene} from 'app/scenes/message/messageScene';
+import {drawFrame} from 'app/utils/animations';
+import {fillRect, pad} from 'app/utils/index';
+import {simpleWhiteFont, keyboardMap, xboxMap} from 'app/utils/simpleWhiteFont';
 
 const font = simpleWhiteFont;
 
@@ -202,7 +203,7 @@ export function renderTextRow(context: CanvasRenderingContext2D, text: string, {
     }
 }
 
-export function renderMessage(context: CanvasRenderingContext2D, state: GameState): void {
+export function renderChoice(context: CanvasRenderingContext2D, state: GameState, scene: MessageScene): void {
     let h = messageRows * (16 + 2) + 6;
     let w = messageWidth + 8;
     let r = {
@@ -215,85 +216,7 @@ export function renderMessage(context: CanvasRenderingContext2D, state: GameStat
     if (state.hero.y - state.camera.y >= 128) {
         r.y = 32;
     }
-    if (state.messagePage) {
-        const rowDuration = 100;
-        fillRect(context, pad(r, 2), 'white');
-        fillRect(context, r, 'black');
-        r = pad(r, -4);
-        // pages[pageIndex] can also be DialogueLootDefinition, but `pageIndex` should never
-        // stop on a loot definition.
-        const startingRow = Math.max(state.messagePage.lineIndex - messageRows, 0);
-        const endingRow = Math.min(state.messagePage.frames.length - 1, state.messagePage.lineIndex + messageRows - 1);
-        //console.log([state.messagePage.lineIndex, state.messagePage.frames.length], [startingRow, endingRow]);
-        //let time = 0;
-        for (let row = startingRow; row <= endingRow; row++) {
-            let x = r.x;
-            const y = (state.messagePage.lineIndex === 0)
-                ? r.y + (row - startingRow) * 18
-                : r.y + (row - startingRow) * 18 - Math.round(18 * Math.max(0, Math.min(endingRow - startingRow - messageRows + 1, state.messagePage.animationTime / rowDuration)));
-            // console.log(state.messagePage.animationTime, row, y - r.y);
-            const frameRow = state.messagePage.frames[row];
-            // This will make one line appear in the first message page at a time every 80ms
-            /*if (state.messagePage.lineIndex === 0 && time >= state.messagePage.animationTime) {
-                break;
-            }
-            time += rowDuration;*/
-            // This will smoothly fill in the first message page
-            // const maxY = state.messagePage.lineIndex === 0 ? r.y + Math.floor(18 * state.messagePage.animationTime / rowDuration) : r.y + r.h;
-            const maxY = r.y + r.h;
-            for (const frame of frameRow) {
-                if (!frame) {
-                    x += characterWidth;
-                    continue;
-                }
-                // Skip lines that are entirely outside of the message box area.
-                if (y + frame.h <= r.y || y >= maxY) {
-                    break;
-                }
-                if (y < r.y) {
-                    const h = r.y - y;
-                    // Only draw the bottom of frames that are scrolling off the top of the message box.
-                    drawFrame(context, {
-                        ...frame,
-                        y: frame.y + h,
-                        h: frame.h - h,
-                    }, {
-                        x: x - (frame.content?.x || 0) * (frame.s || 1),
-                        y: y - (frame.content?.y || 0) * (frame.s || 1) + h,
-                        w: frame.w * (frame.s || 1),
-                        h: frame.h * (frame.s || 1) - h,
-                    });
-                } else if (y + frame.h > maxY) {
-                    const h = (y + frame.h) - maxY;
-                    // Only draw the top of frames that are scrolling off the top of the message box.
-                    drawFrame(context, {
-                        ...frame,
-                        h: frame.h - h,
-                    }, {
-                        x: x - (frame.content?.x || 0) * (frame.s || 1),
-                        y: y - (frame.content?.y || 0) * (frame.s || 1),
-                        w: frame.w * (frame.s || 1),
-                        h: frame.h * (frame.s || 1) - h,
-                    });
-                } else {
-                    drawFrame(context, frame, {
-                        x: x - (frame.content?.x || 0) * (frame.s || 1),
-                        y: y - (frame.content?.y || 0) * (frame.s || 1),
-                        w: frame.w * (frame.s || 1),
-                        h: frame.h * (frame.s || 1),
-                    });
-                }
-                x += frame.w * (frame.s || 1);
-                /*time += charTime;
-                if (state.messagePage.lineIndex === 0 && time > state.messagePage.animationTime) {
-                    break;
-                }*/
-            }
-            /*if (state.messagePage.lineIndex === 0 && time > state.messagePage.animationTime) {
-                break;
-            }*/
-        }
-    }
+
     const choice = state.scriptEvents.activeEvents.find(event => event.type === 'showChoiceBox') as ShowChoiceBoxActiveScriptEvent;
     if (choice) {
         let choiceIndex = choice.choiceIndex || 0;
