+++ app/scenes/scenes.ts
@@ -1,6 +1,8 @@
 export * from 'app/scenes/field/fieldScene';
 export * from 'app/scenes/fieldMenu/fieldMenuScene';
 export * from 'app/scenes/fileSelect/fileSelectScene';
+export * from 'app/scenes/hud/hudScene';
 export * from 'app/scenes/intro/introScene';
 export * from 'app/scenes/map/mapScene';
+export * from 'app/scenes/pause/pauseScene';
 export * from 'app/scenes/title/titleScene';
